requirements.txt
Django~=3.2.18
djangorestframework~=3.14.0
psycopg2==2.9.5
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-trench==0.3.1
djangorestframework-simplejwt==5.2.2
django-extensions==3.2.1